// src/web/src/app/api/bff/models/institution-created.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InstitutionInvitationCreated } from '../models/institution-invitation-created';
import { InstitutionSummary } from '../models/institution-summary';
export interface InstitutionCreated {
  additionalData?: ({
} | null);
  institution?: (null | InstitutionSummary);
  invitation?: (null | InstitutionInvitationCreated);
}