// src/web/src/app/api/bff/models/institution-invitation-created.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface InstitutionInvitationCreated {
  additionalData?: ({
} | null);
  email?: (string | null);
  expiresAt?: (string | null);
  id?: (string | null);
  institutionId?: (string | null);
  invitationUrl?: (string | null);
  role?: (string | null);
}